+++ src/main/java/ru/kata/spring/boot_security/demo/dao/RoleDao.java
@@ -5,11 +5,11 @@ import ru.kata.spring.boot_security.demo.models.Role;
 import java.util.List;
 
 public interface RoleDao {
-    Role getByIdRole(Long id);
+    Role getByIdRole(int id);
     List<Role> getListRoles();
     Role getByName(String name);
     List<Role> getListByName(List<String> name);
     boolean add(Role user);
 
-    List<Role> getRolesListById(List<Long> roles);
+    List<Role> getRolesListById(List<Integer> roles);
 }
