+++ src/main/java/ru/kata/spring/boot_security/demo/service/RoleService.java
@@ -10,8 +10,8 @@ public interface RoleService {
 
     List<Role> getListRoles();
 
-    List<Role> getRolesListById(List<Long> id);
+    List<Role> getRolesListById(List<Integer> id);
 
-    Role getRoleById(Long id);
+    Role getRoleById(int id);
 
 }
