+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserServiceImpl.java
@@ -57,7 +57,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public void delete(Long id) {
+    public void delete(int id) {
         userDao.delete(id);
     }
 
@@ -74,7 +74,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User getById(Long id) {
+    public User getById(int id) {
         return userDao.getById(id);
     }
